+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Changes in version 0.1.1
 - Add the `beers` data and source file <2019-02-11 Mon>
 - Add the `clinical` data and source file <2019-02-11 Mon>
+- Add a function to get the package version <2019-02-12 Tue>
 
 ## Changes in version 0.1.0